requirements.txt
Flask==2.2.2
pyyaml==6.0
requests==2.28.2
